requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
psutil==5.9.6
requests==2.31.0
customtkinter==5.2.0
Pillow==10.0.1 